+++ packages/stream_chat/test/src/core/api/general_api_test.dart
@@ -7,6 +7,7 @@ import 'package:stream_chat/stream_chat.dart';
 import 'package:test/test.dart';
 
 import '../../mocks.dart';
+import '../../utils.dart';
 
 void main() {
   Response successResponse(String path, {Object? data}) => Response(
@@ -310,6 +311,24 @@ void main() {
         queryParameters: {'url': url},
       ),
     ).called(1);
+    verifyNoMoreInteractions(client);
+  });
+
+  test('getAppSettings', () async {
+    const path = '/app';
+    final fixture = jsonFixture('app_settings.json');
+
+    when(() => client.get(path)).thenAnswer(
+      (_) async => successResponse(path, data: fixture),
+    );
+
+    final res = await generalApi.getAppSettings();
+
+    expect(res.app.name, 'test-app');
+    expect(res.app.fileUploadConfig.blockedFileExtensions, ['.exe']);
+    expect(res.app.imageUploadConfig.sizeLimit, 5242880);
+
+    verify(() => client.get(path)).called(1);
     verifyNoMoreInteractions(client);
   });
 }