+++ packages/stream_chat_localizations/example/lib/add_new_lang.dart
@@ -161,6 +161,12 @@ class NnStreamChatLocalizations extends GlobalStreamChatLocalizations {
       'The file is too large to upload. The file size limit is $limitInMB MB.';
 
   @override
+  String fileTypeNotSupportedError(String? extension) {
+    if (extension == null) return 'This file type is not supported for upload.';
+    return "'.$extension' files are not supported for upload.";
+  }
+
+  @override
   String get couldNotReadBytesFromFileError =>
       'Could not read bytes from file.';